productivity_tools_I/requirements.txt
Flask==3.1.1
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
Flask-Security-Too==5.4.3
Flask-Migrate==4.0.7
Flask-Lucide==0.2.0
Flask-WTF==1.2.1
WTForms==3.1.2
email-validator==2.1.1
bcrypt==4.3.0

